aggiustata tabella followers

+++ service/database/unfollow-user.go
@@ -14,10 +14,20 @@ func (db *appdbimpl) UnfollowUser(id1 int, id2 int) (string, error) {
 		return "", err
 	}
 
+	var count3 int
+	err = db.c.QueryRow("SELECT COUNT(*) FROM follow WHERE id1 = ? AND id2=?", id1, id2).Scan(&count3)
+	if err != nil {
+		return "", err
+	}
+
 	if count1 > 0 && count2 > 0 {
-		_, err := db.c.Exec("DELETE FROM follow WHERE id1=? AND id2=?", id1, id2)
-		if err != nil {
-			return "", err
+		if count3 > 0 {
+			_, err := db.c.Exec("DELETE FROM follow WHERE id1=? AND id2=?", id1, id2)
+			if err != nil {
+				return "", err
+			}
+		} else {
+			return "utente non presente nella tabella follow", nil
 		}
 	} else {
 		return "utente non presente", nil
